gcloud/requirements.txt
numpy==1.17.2
requests==2.18.4
https://download.pytorch.org/whl/cpu/torch-1.0.1.post2-cp37-cp37m-linux_x86_64.whl
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-core==1.0.3
google-cloud-pubsub==1.0.2
google-cloud-storage==1.20